Fix major bug in song data composition.
The delta times for MIDI events should be relative times between
successive events. This commit inclues an in-progress unit test for
datasets that ensures that a song loaded, rasterized, and then composed
will contain the same data. Presently, this is not possible for many
songs because start-stop sequences of the same note are lost in the
rasterized format.